+++ src/dataStructures/tree/binarySearchTree/utils/traverseAfterOrder.ts
@@ -7,14 +7,13 @@ export function traverseAfterOrder<T = unknown>(
 ): void {
   const nodeStack = new Stack<BinarySearchTreeNodeInterface<T>>();
   let currentNode = root;
-  let peekNode;
   if (root) {
     nodeStack.push(root);
   }
 
   while (!nodeStack.isEmpty()) {
     if (nodeStack.peek() !== currentNode.parent) {
-      peekNode = nodeStack.peek();
+      let peekNode = nodeStack.peek();
       while (peekNode.left || peekNode.right) {
         if (peekNode.left && peekNode.right) {
           nodeStack.push(peekNode.right);