+++ src/dataStructures/tree/binarySearchTree/utils/traverseInOrder.ts
@@ -7,14 +7,13 @@ export function traverseInOrder<T = unknown>(
 ): void {
   const nodeStack = new Stack<BinarySearchTreeNodeInterface<T>>();
   let currentNode = root;
-  let tmpNode;
 
   while (currentNode || !nodeStack.isEmpty()) {
     if (currentNode) {
       nodeStack.push(currentNode);
       currentNode = currentNode.left;
     } else {
-      tmpNode = nodeStack.pop();
+      const tmpNode = nodeStack.pop();
       if (callback(tmpNode) === false) {
         return;
       }