+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 0.0.0 (2021-04-23)
+# 0.0.0 (2021-04-24)
 
 ### Bug Fixes
 
@@ -62,10 +62,11 @@
 - **BST:** 增加先序、后序、层序遍历 ([488b4a8](https://github.com/air-supply94/js-algorithm/commit/488b4a8408fd8a40aec420f0ce8a32b3b17c9bce))
 - **BST:** 增加后序遍历的迭代版 ([48a8015](https://github.com/air-supply94/js-algorithm/commit/48a80158438dadb469d72bbb3f5bde0deab584ee))
 - **bst/leetcode:** add connectNodeNext ([1115283](https://github.com/air-supply94/js-algorithm/commit/11152831b8cdb1f17ad70e867267c5062c2bab53))
-- **bst/leetcode:** add constructMaximumTree ([3830552](https://github.com/air-supply94/js-algorithm/commit/38305525c4780005141ceeb936ecfb68b6ff0d4b))
+- **bst/leetcode:** add constructMaximumTree ([a342aaf](https://github.com/air-supply94/js-algorithm/commit/a342aaf0668ccd6c9b8f859e05e2107c9e0529e3))
 - **bst/leetcode:** add flattenToLinkedList ([eb61761](https://github.com/air-supply94/js-algorithm/commit/eb617614ab93723a000004e5f7b462fa89d96272))
 - **bst/leetcode:** add invertTree ([d06b1ff](https://github.com/air-supply94/js-algorithm/commit/d06b1ffd483d99e332e2fa31ad207c375cede0c6))
 - **bst/leetcode:** add isSameMetric ([ba0a053](https://github.com/air-supply94/js-algorithm/commit/ba0a0536fd76b4d069c254857b383939341f716e))
+- **bst/leetcode:** add serializePreAndInOrder ([e766011](https://github.com/air-supply94/js-algorithm/commit/e766011b8fedd3901da675a21cbc84940317471e))
 - **Comparator:** 更新导出方式和访问限制 ([ef485dd](https://github.com/air-supply94/js-algorithm/commit/ef485ddc67e6743f500e1ab0e8cec6f876ad1555))
 - **count-sort:** add ([c1d588d](https://github.com/air-supply94/js-algorithm/commit/c1d588d6de86c0f60e0385e56bdb9c71bc1b1a88))
 - **dataStructure:** add bitMap ([252f1b6](https://github.com/air-supply94/js-algorithm/commit/252f1b6f4901f96725b9b764e55237ab4bae0303))